% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dislgum.R
\name{dislgum}
\alias{dislgum}
\title{Log gumbel distribution function for 24-hour maximum precipitation}
\usage{
dislgum(x, tr = NULL)
}
\arguments{
\item{x}{Vector or dataframe of daily maximum precipitation values}

\item{tr}{OPTIONAL, Vector of return period values, default parameter is NULL.
Typical values for TR are 2, 5, 10, 20, 50, 100, 200, 500, 1000 years.}
}
\description{
Log gumbel distribution function for 24-hour maximum precipitation for the study
of extreme precipitation events.
}
\author{
Geomar Perales Apaico
}
